+++ galacteek/ui/themes/breeze-dark/__init__.py
@@ -1,0 +1,5 @@
+from galacteek.ui.themes import Theme
+
+
+theme = Theme(name='breeze-dark')
+theme.colors.webEngineBackground = '#31363b'