+++ galacteek/ui/themes/indigo/__init__.py
@@ -1,0 +1,5 @@
+from galacteek.ui.themes import Theme
+
+
+theme = Theme(name='indigo')
+theme.colors.webEngineBackground = '#14141E'